Big update booom

+++ src/GameObject.cs
@@ -4,18 +4,18 @@ namespace SharpEngine;
 
 public class GameObject
 {
-	public int X {get; private set;}
-	public int Y {get; private set;}
+	public double X {get; private set;}
+	public double Y {get; private set;}
 	public char Char {get; private set;}
 
-	public GameObject(char _char, int _x = 0, int _y = 0)
+	public GameObject(char _char, double _x = 0.0d, double _y = 0.0d)
 	{
 		X = _x;
 		Y = _y;
 		Char = _char;
 	}
 
-	public void position(int _x, int _y)
+	public void position(double _x, double _y)
 	{
 		X = _x;
 		Y = _y;
